+++ metadata.rb
@@ -5,7 +5,7 @@ license          'All rights reserved'
 description      'Installs/Configures rackops_rolebook'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 
-version          '1.3.1'
+version          '1.4.0'
 
 depends 'chef-client',        '~> 3.3'
 depends 'rackspace_user',     '~> 1.0'
@@ -15,3 +15,4 @@ depends 'rackspace_ntp',      '~> 2.0'
 depends 'rackspace_iptables', '~> 1.2'
 depends 'rackspace_openssh',  '~> 3.1'
 depends 'rackspace_cloudmonitoring', '~> 3.1'
+depends 'ohai', '~> 1.1'