+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file is used to list changes made in each version of rackops_rolebook.
 
+## 1.4.0
+* Add public_info Ohai plugin to populate public_info attributes (https://github.com/rackops/rackops_rolebook/pull/14)
+
+## 1.3.1
+* Add Openssh dependency (https://github.com/rackops/rackops_rolebook/pull/11)
+
+## 1.3.0
+* Add Logrotation (https://github.com/rackops/rackops_rolebook/pull/9)
+
+## 1.4.0
+* Add public_info Ohai plugin to populate public_info attributes
+
 ## 1.2.0
 * Add rackspace_cloudmonitoring as a dependency to enable standard checks